requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.13
Flask==3.1.1
flask-cors==6.0.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
imap-tools==1.11.0
imaplib2==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
Werkzeug==3.1.3
